+++ src/app/[locale]/(main-public)/_components/LandingPage/AdmissionOpen.tsx
@@ -1,4 +1,4 @@
-import AdmissionImage from "@/../public/images/Screenshot_5.png";
+import AdmissionBannerImage from "@/../public/images/landing-page-admission-banner.png";
 import { getScopedI18n } from "@/locales/server";
 import { Button } from "@lib/next-ui";
 import NextImage from "next/image";
@@ -32,7 +32,7 @@ const AdmissionOpen: FC = async () => {
         width={500}
         height={500}
         quality={95}
-        src={AdmissionImage}
+        src={AdmissionBannerImage}
         alt="Students Studying"
         className="order-4 row-span-2 ml-auto hidden max-w-xs rounded-lg object-cover object-center shadow-2xl lg:block"
       />
