+++ src/app/[locale]/(main-public)/_components/LandingPage/HeroCarousel.tsx
@@ -2,18 +2,18 @@
 
 import { BiLeftArrow, BiRightArrow } from "react-icons/bi";
 
-import MainCarouselImage2 from "@/../public/images/Screenshot_1.png";
-import MainCarouselImage3 from "@/../public/images/Screenshot_2.png";
-import MainCarouselImage4 from "@/../public/images/Screenshot_3.png";
-import MainCarouselImage from "@/../public/images/mainCarosoulImg.png";
-import MainCarouselImage1 from "@/../public/images/mainCarosoulImg1.png";
+import MainCarouselImage1 from "@/../public/images/hero-carousel/1.png";
+import MainCarouselImage2 from "@/../public/images/hero-carousel/2.png";
+import MainCarouselImage3 from "@/../public/images/hero-carousel/3.png";
+import MainCarouselImage4 from "@/../public/images/hero-carousel/4.png";
+import MainCarouselImage5 from "@/../public/images/hero-carousel/5.png";
 import Carousel from "@/components/ui/Carousel";
 import NextImage from "next/image";
 import type { FC } from "react";
 import { v4 as uuid } from "uuid";
 
 const imagesLinks = [
-  MainCarouselImage,
+  MainCarouselImage5,
   MainCarouselImage1,
   MainCarouselImage2,
   MainCarouselImage3,
